+++ src/scipp/typing.py
@@ -5,7 +5,10 @@
 
 import typing as _std_typing
 
+import numpy.typing
+
 from ._scipp import core as sc
+from .core.cpp_classes import DataArray, Dataset, DType, Variable
 
 
 def is_scalar(obj: _std_typing.Any) -> bool:
@@ -43,14 +46,36 @@ def has_numeric_type(obj: _std_typing.Any) -> bool:
     return (not has_vector_type(obj)) and (not has_string_type(obj))
 
 
-#: Any object that behaves like a scipp.Variable,
-#:  that is an array with labeled dimensions.
-VariableLike = _std_typing.Union[sc.Variable, sc.DataArray, sc.Dataset]
+VariableLike = _std_typing.Union[Variable, DataArray, Dataset]
+"""Any object that behaves like a :class:`scipp.Variable`.
 
-#: dict-like object mapping dimension labels to Variables.
-MetaDataMap = _std_typing.MutableMapping[str, sc.Variable]
+More concretely, an array with labeled dimensions:
 
-# TypeVar for use in annotations.
-# Should be hidden in rendered documentation in favor of VariableLike.
-VariableLikeType = _std_typing.TypeVar('VariableLikeType', sc.Variable, sc.DataArray,
-                                       sc.Dataset)
+- :class:`scipp.DataArray`
+- :class:`scipp.Dataset`
+- :class:`scipp.Variable`
+"""
+
+MetaDataMap = _std_typing.MutableMapping[str, Variable]
+"""dict-like object mapping dimension labels to Variables."""
+
+VariableLikeType = _std_typing.TypeVar('VariableLikeType', Variable, DataArray, Dataset)
+"""TypeVar for use in annotations.
+
+Should be hidden in rendered documentation in favor of VariableLike.
+"""
+
+DTypeLike = _std_typing.Union[numpy.typing.DTypeLike, DType]
+"""Anything that can be interpreted as a dtype.
+
+This includes
+
+- :class:`scipp.DType`
+- everything that is supported by
+  `numpy.DTypeLike <https://numpy.org/devdocs/reference/typing.html#numpy.typing.DTypeLike>`_
+  e.g.
+
+  - :class:`numpy.dtype`
+  - :class:`type` objects like :class:`int` and :class:`float`
+  - names of dtypes as strings like ``'int32'`` and ``'float64'``
+"""  # noqa: E501